gltf: Example Simplifications
