+++ examples/gltf/package.json
@@ -9,11 +9,11 @@
     "open-vr": "adb reverse tcp:8080 tcp:8080 && adb shell am start -a android.intent.action.VIEW -d http://localhost:8080/"
   },
   "dependencies": {
-    "@loaders.gl/core": "^1.2.0",
-    "@loaders.gl/draco": "^1.2.0",
-    "@luma.gl/addons": "^7.3.0-alpha.6",
-    "@luma.gl/constants": "^7.3.0-alpha.6",
-    "@luma.gl/core": "^7.3.0-alpha.6",
+    "@loaders.gl/core": "^1.3.0-beta.1",
+    "@loaders.gl/draco": "^1.3.0-beta.1",
+    "@luma.gl/addons": "^7.3.0-alpha.7",
+    "@luma.gl/constants": "^7.3.0-alpha.7",
+    "@luma.gl/core": "^7.3.0-alpha.7",
     "math.gl": "^3.0.0-beta.3"
   },
   "devDependencies": {
